refactor(useisabovebreakpoint): update tests to follow new standards
